structural refactoring
