+++ commands/anime/anime-search.js
