+++ commands/personal/user.js
@@ -1,6 +1,6 @@
 const { SlashCommandBuilder } = require('discord.js');
 const { userMention } = require('@discordjs/builders');
-const { wife } = require('../config.json');
+const { wife } = require('../../config.json');
 
 module.exports = {
 	data: new SlashCommandBuilder()
